requirements.txt
astroid==2.0.4
autopep8==1.4
certifi==2018.8.24
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.1.2
django-debug-toolbar==1.10.1
djangorestframework==3.8.2
drf-yasg==1.10.2
idna==2.7
inflection==0.3.1
isort==4.3.4
itypes==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
psycopg2==2.7.5
pycodestyle==2.4.0
pylint==2.1.1
python-decouple==3.1
pytz==2018.5
requests==2.19.1
ruamel.yaml==0.15.72
six==1.11.0
sqlparse==0.2.4
uritemplate==3.0.0
urllib3==1.23
wrapt==1.10.11
